imweb/im-backend/fastapi-backend/requirements.txt
FastAPI
uvicorn
pydantic
requests
python-dotenv